agent/requirements.txt
flask
flask_cors
requests
anthropic
feedparser
bs4
ollama
chromadb
sentence-transformers
python-dotenv